check that amis exist (#34)
This should be everything we need.  We could probably check for a couple other things we commonly have: security-groups and kernel ids.  We could probably also check for the AMI in the provisioning iterations and skip worker types which have any missing AMIs, security groups or kernels, or at least the region that has those missing.