Use key name to declare ownership, but use a static key name
this addresses our security concern since we will no longer let the provisioner manage the keys that it uses.  This means that getting the provisioner's aws credentials will no longer allow someone to specify the root ssh key for the AMIs that are accessible.  That means that you'd need a different set of aws credentials than the one in the Heroku account or the Provisioner configs