Use EC2-Manager state directly.
This also removes the scopes from the /state/ method because that information really isn't confidential.  This method wraps the EC2-Manager state and stats methods to provide a compatibility path forward until the tools website can be updated.  NOTE: The /state/ method's summary property is derived from a different call to the ec2-manager than the one to generate the lists of instances and requests.  This is done because it's non-trivial to use the result of the state call to ec2-manager in the summary generation function