when we do not have a capacity for an instance type, assume 1
This is an issue when we have running instance which are of an instance type
that is no longer included in the worker type definition.  In this case, we're
choosing to set the capacity to 1.  The assumption is that if the instance
would no longer be valid that the worker should shut itself down upon receiving
the updated worker type message.  We can also assume that capacity is at least
one because any lower wouldn't be possible.  Assuming a capacity of 1 also
ensures that we count running old instance for some value so that we don't
double up too much on resources