Update mocha (adding --exit)
In version 4.0.0 of mocha, the --exit flag was added because of a change
to default behaviour.  The change is that mocha will no longer exit
after the tests complete.  Ideally, this wouldn't require using the
--exit flag in our tests, but something is keeping it alive.  It's
probably a server or some listener, but for the time being, let's update
the mocha dependency while using the pre-4.0.0
process.exit()-on-test-complete behaviour